+++ src/app/components/TopBar/container.js
@@ -2,8 +2,8 @@ import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 
 import TopBar from "./TopBar";
-import { toggleDrawer } from "../SideDrawer/actions";
-import { handleCalendarChange } from "../DailyForm/actions";
+import { toggleDrawer } from "../../modules/SideDrawer/actions";
+import { handleCalendarChange } from "../../modules/DailyForm/actions";
 
 const mapStateToProps = state => ({
   today: state.today.today
