+++ src/app/modules/DailyForm/DailyForm.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 import { startOfMinute, addDays } from "date-fns";
-import { Redirect } from "react-router-dom";
 
 import Card from "@material-ui/core/Card";
 import LinearProgress from "@material-ui/core/LinearProgress";
@@ -9,7 +8,7 @@ import Button from "../../components/Button";
 
 import styled from "styled-components";
 
-import TopBar from "../TopBar";
+import TopBar from "../../components/TopBar";
 
 import TopControls from "./TopControls";
